% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_shiny.R
\name{annotate_shiny}
\alias{annotate_shiny}
\title{Run the annotation pipeline on a list of variants from shiny app}
\usage{
annotate_shiny(config.list)
}
\arguments{
\item{config.list}{List. A list of variants andconfiguration parameters.}
}
\value{
a data table with all variant information is returned.
}
\description{
This function should not be used outside shiny app
A list of variants and parameters are received and their information is checked on various API servers.
}
