% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_annotation}
\alias{demo_annotation}
\title{Demo run of the annotation pipeline}
\usage{
demo_annotation()
}
\value{
A data table containing all variant information is returned based on the user's selected
specifications and parameters. Additionally, numerous other report files in various formats,
including text, HTML, Excel, and image, are saved in the output folder.
}
\description{
This function is a demo of the annotation algorithm.
}
