% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_funcs.R
\name{EnsemblReleases}
\alias{EnsemblReleases}
\title{data release available on this REST server.}
\usage{
EnsemblReleases(server = "https://rest.ensembl.org")
}
\arguments{
\item{server}{Ensemble API end point. default: https://rest.ensembl.org}
}
\value{
a message is displayed to the user
}
\description{
Shows the data releases available on this REST server.
May return more than one release (infrequent non-standard Ensembl configuration).
}
