% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configFile_funcs.R
\name{getConfigFile}
\alias{getConfigFile}
\title{Copy a sample configuration file}
\usage{
getConfigFile(dir.path)
}
\arguments{
\item{dir.path}{The existing folder for copying the file.}
}
\description{
This function provides a sample configuration file. The user can modify the parameters as desired
}
