% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_funcs.R
\name{getEBIgroups}
\alias{getEBIgroups}
\title{Get the list of available tissue data from EBI website.}
\usage{
getEBIgroups()
}
\arguments{
\item{server}{the address for EBI API server.}
}
\value{
A data.table containing the results.
}
\description{
This function can be used to check the available tissue data from EBI API server.
}
