% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_files.R
\name{mergeResultFiles}
\alias{mergeResultFiles}
\title{Merge multiple output files}
\usage{
mergeResultFiles(..., fileName)
}
\arguments{
\item{...}{list of input files to be merged.}

\item{fileName}{name of the output file.}
}
\value{
A data table is returned.
}
\description{
This function merges multiple result files into one.
}
