% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{releaseVersion}
\alias{releaseVersion}
\title{Shows the data releases available on this REST server.}
\usage{
releaseVersion(server)
}
\arguments{
\item{server}{name of the server.}
}
\value{
a message is displayed to the user
}
\description{
Shows the data releases available on this REST server.
May return more than one release (unfrequent non-standard Ensembl configuration).
}
