% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineOnlist.R
\name{run_annotation}
\alias{run_annotation}
\title{Run the annotation pipeline on a list of variants}
\usage{
run_annotation(configurationFilePath, verbose = TRUE)
}
\arguments{
\item{configurationFilePath}{Character. The path to the configuration file.}

\item{verbose}{Logical. Whether to display messages in the console.}
}
\value{
A data table containing all variant information is returned based on the user's selected
specifications and parameters. Report files in various formats, including text, HTML,
Excel, and image, are saved in the output folder.
}
\description{
This function receives the path to the configuration file.
A list of variants is received and their information is checked on various API servers.
}
