% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_annotation}
\alias{demo_annotation}
\title{Demo run of the annotation pipeline}
\usage{
demo_annotation()
}
\value{
A data table containing the variant information for testing is returned.
Report files are also saved in the current working directory.
}
\description{
This function is a demo of the annotation algorithm.
}
