% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_funcs.R
\name{EnsemblDatabases}
\alias{EnsemblDatabases}
\title{List population from human database (1000 Genomes project)}
\usage{
EnsemblDatabases(build = 38)
}
\arguments{
\item{build}{Genome build. Either 37 or 38. default: 38}
}
\value{
A data table is returned which includes the name, description and size of the available populations
in 1000 Genomes project database.
}
\description{
This function list the name, description and size of the available populations
in 1000 Genomes project database. This database will be used for returning variables in high LD
with the target SNP.
}
