% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_funcs.R
\name{EnsemblReleases}
\alias{EnsemblReleases}
\title{data release available on this REST server.}
\usage{
EnsemblReleases(build = 38)
}
\arguments{
\item{build}{Genome build. Either 37 or 38. default: 38}
}
\value{
a message is displayed to the user
}
\description{
Shows the data releases available on this REST server.
May return more than one release (infrequent non-standard Ensembl configuration).
}
